clean up some immutable deserialized packet tests
