+++ core/src/unprocessed_packet_batches.rs
@@ -3,16 +3,11 @@ use {
     min_max_heap::MinMaxHeap,
     solana_perf::packet::{Packet, PacketBatch},
     solana_runtime::transaction_priority_details::TransactionPriorityDetails,
-    solana_sdk::{
-        feature_set,
-        hash::Hash,
-        transaction::{AddressLoader, SanitizedTransaction, Transaction},
-    },
+    solana_sdk::{hash::Hash, transaction::Transaction},
     std::{
         cmp::Ordering,
         collections::{hash_map::Entry, HashMap},
         rc::Rc,
-        sync::Arc,
     },
 };
 
@@ -322,43 +317,11 @@ pub fn transactions_to_deserialized_packets(
         })
         .collect()
 }
-
-// This function deserializes packets into transactions, computes the blake3 hash of transaction
-// messages, and verifies secp256k1 instructions. A list of sanitized transactions are returned
-// with their packet indexes.
-#[allow(clippy::needless_collect)]
-pub fn transaction_from_deserialized_packet(
-    deserialized_packet: &ImmutableDeserializedPacket,
-    feature_set: &Arc<feature_set::FeatureSet>,
-    votes_only: bool,
-    address_loader: impl AddressLoader,
-) -> Option<SanitizedTransaction> {
-    if votes_only && !deserialized_packet.is_simple_vote() {
-        return None;
-    }
-
-    let tx = SanitizedTransaction::try_new(
-        deserialized_packet.transaction().clone(),
-        *deserialized_packet.message_hash(),
-        deserialized_packet.is_simple_vote(),
-        address_loader,
-    )
-    .ok()?;
-    tx.verify_precompiles(feature_set).ok()?;
-    Some(tx)
-}
-
 #[cfg(test)]
 mod tests {
     use {
         super::*,
-        solana_perf::packet::PacketFlags,
-        solana_sdk::{
-            signature::{Keypair, Signer},
-            system_transaction,
-            transaction::{SimpleAddressLoader, Transaction},
-        },
-        solana_vote_program::vote_transaction,
+        solana_sdk::{signature::Keypair, system_transaction},
     };
 
     fn simple_deserialized_packet() -> DeserializedPacket {
@@ -485,133 +448,5 @@ mod tests {
         );
         assert!(unprocessed_packet_batches.is_empty());
         assert!(unprocessed_packet_batches.pop_max_n(0).is_none());
-    }
-
-    #[cfg(test)]
-    fn make_test_packets(
-        transactions: Vec<Transaction>,
-        vote_indexes: Vec<usize>,
-    ) -> Vec<DeserializedPacket> {
-        let capacity = transactions.len();
-        let mut packet_vector = Vec::with_capacity(capacity);
-        for tx in transactions.iter() {
-            packet_vector.push(Packet::from_data(None, &tx).unwrap());
-        }
-        for index in vote_indexes.iter() {
-            packet_vector[*index].meta.flags |= PacketFlags::SIMPLE_VOTE_TX;
-        }
-
-        packet_vector
-            .into_iter()
-            .map(|p| DeserializedPacket::new(p).unwrap())
-            .collect()
-    }
-
-    #[test]
-    fn test_transaction_from_deserialized_packet() {
-        use solana_sdk::feature_set::FeatureSet;
-        let keypair = Keypair::new();
-        let transfer_tx =
-            system_transaction::transfer(&keypair, &keypair.pubkey(), 1, Hash::default());
-        let vote_tx = vote_transaction::new_vote_transaction(
-            vec![42],
-            Hash::default(),
-            Hash::default(),
-            &keypair,
-            &keypair,
-            &keypair,
-            None,
-        );
-
-        // packets with no votes
-        {
-            let vote_indexes = vec![];
-            let packet_vector =
-                make_test_packets(vec![transfer_tx.clone(), transfer_tx.clone()], vote_indexes);
-
-            let mut votes_only = false;
-            let txs = packet_vector.iter().filter_map(|tx| {
-                transaction_from_deserialized_packet(
-                    tx.immutable_section(),
-                    &Arc::new(FeatureSet::default()),
-                    votes_only,
-                    SimpleAddressLoader::Disabled,
-                )
-            });
-            assert_eq!(2, txs.count());
-
-            votes_only = true;
-            let txs = packet_vector.iter().filter_map(|tx| {
-                transaction_from_deserialized_packet(
-                    tx.immutable_section(),
-                    &Arc::new(FeatureSet::default()),
-                    votes_only,
-                    SimpleAddressLoader::Disabled,
-                )
-            });
-            assert_eq!(0, txs.count());
-        }
-
-        // packets with some votes
-        {
-            let vote_indexes = vec![0, 2];
-            let packet_vector = make_test_packets(
-                vec![vote_tx.clone(), transfer_tx, vote_tx.clone()],
-                vote_indexes,
-            );
-
-            let mut votes_only = false;
-            let txs = packet_vector.iter().filter_map(|tx| {
-                transaction_from_deserialized_packet(
-                    tx.immutable_section(),
-                    &Arc::new(FeatureSet::default()),
-                    votes_only,
-                    SimpleAddressLoader::Disabled,
-                )
-            });
-            assert_eq!(3, txs.count());
-
-            votes_only = true;
-            let txs = packet_vector.iter().filter_map(|tx| {
-                transaction_from_deserialized_packet(
-                    tx.immutable_section(),
-                    &Arc::new(FeatureSet::default()),
-                    votes_only,
-                    SimpleAddressLoader::Disabled,
-                )
-            });
-            assert_eq!(2, txs.count());
-        }
-
-        // packets with all votes
-        {
-            let vote_indexes = vec![0, 1, 2];
-            let packet_vector = make_test_packets(
-                vec![vote_tx.clone(), vote_tx.clone(), vote_tx],
-                vote_indexes,
-            );
-
-            let mut votes_only = false;
-            let txs = packet_vector.iter().filter_map(|tx| {
-                transaction_from_deserialized_packet(
-                    tx.immutable_section(),
-                    &Arc::new(FeatureSet::default()),
-                    votes_only,
-                    SimpleAddressLoader::Disabled,
-                )
-            });
-            assert_eq!(3, txs.count());
-
-            votes_only = true;
-            let txs = packet_vector.iter().filter_map(|tx| {
-                transaction_from_deserialized_packet(
-                    tx.immutable_section(),
-                    &Arc::new(FeatureSet::default()),
-                    votes_only,
-                    SimpleAddressLoader::Disabled,
-                )
-            });
-            assert_eq!(3, txs.count());
-        }
     }
 }
